unpack-api/pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "unpack_api"
requires-python = ">=3.10"
version = '0.0.1'
authors = [
    { name = "Max Fatouros", email = "max.fatouros@psi.ch" }
]
# description = "description"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "authlib",
    "fastapi[standard]",
    "python-dotenv",
    "requests",
]

[project.optional-dependencies]
dev = ['pre-commit']


[tool.setuptools]
py-modules = [
    "unpacked_api",
]

# [project.urls]
